requirements.txt
langchain
langchain-openai
langchain-ollama
langchain-community
langchainhub
python-dotenv
black
tweepy
streamlit